Fixed data leakage
I changed the ScanSlice to have a case value,
The MRID to have 4 new functions to facilitate better splitting of the data, one of which lets you approximate a percentage of the total number of slices and will tell you which cases you need to accomplish that.
Then I changed the validation and test loaders to instead use a basic set of data, while the training will still be split according to full, mixed, and empty as before.